+++ app/views/index.blade.php
@@ -128,13 +128,15 @@
                                         return confirm("Are you sure you want to delete the network {{ $network->name }}?");
                                     }
                                 </script>
-                                {{ Form::open(array('action' => array('NetworkController@deleteNetwork', $network->id), 'class' => 'form-horizontal', 'method'=>'DELETE', 'onsubmit' => 'return ConfirmDelete()')) }}
-                                    <div style="margin-top:10px" class="form-group">
-                                        <div class="col-md-12">
-                                            {{ Form::submit('Delete', array('class'=>'btn btn-danger')) }}
+                                @if(Auth::user()->can('delete_network'))
+                                    {{ Form::open(array('action' => array('NetworkController@deleteNetwork', $network->id), 'class' => 'form-horizontal', 'method'=>'DELETE', 'onsubmit' => 'return ConfirmDelete()')) }}
+                                        <div style="margin-top:10px" class="form-group">
+                                            <div class="col-md-12">
+                                                {{ Form::submit('Delete', array('class'=>'btn btn-danger')) }}
+                                            </div>
                                         </div>
-                                    </div>
-                                {{ Form::close() }}
+                                    {{ Form::close() }}
+                                @endif
                             </div>
                         </div>
                     </div>
